Updated User Interface and Aesthetics
This update to the original Visitorials website brings a new minimal look and feel to the website. The main changes are 1) the implementation of the Skeleton CSS boilerplate for browser responsiveness, and 2) the table of entries changed to individualized entry "cards". Development is still in progress but I wanted to show off what I've accomplished thus far.